Refactor dbscan to typescript
