+++ packages/turf-clusters-dbscan/.gitignore
@@ -1,0 +1,1 @@
+index.js
